chore: rename "TracerMiddleware" to "TracerPlugin"
